backend/requirements.txt
fastapi
uvicorn
asyncpg
PyJWT
bcrypt
httpx
google-cloud-storage
google-cloud-bigquery
google-auth
jinja2
psycopg2-binary
python-dotenv
pandas
sqlalchemy
requests
pymysql
pymssql
pyathena
duckdb
duckdb-engine
